Move .empty tests to separate file.
