+++ test/rx/core/test_observable_creation.rb
@@ -5,44 +5,6 @@ require "#{File.dirname(__FILE__)}/../../test_helper"
 class TestObservableCreation < Minitest::Test
   include Rx::ReactiveTest
 
-  # Empty methods
-
-  def test_empty_basic
-    scheduler = Rx::TestScheduler.new
-
-    res = scheduler.configure do
-      Rx::Observable.empty(scheduler)
-    end
-
-    msgs = [on_completed(201)]
-    assert_messages msgs, res.messages       
-  end
-
-  def test_empty_disposed
-    scheduler = Rx::TestScheduler.new
-
-    res = scheduler.configure({:disposed => 200}) do
-      Rx::Observable.empty(scheduler)
-    end
-
-    msgs = []
-    assert_messages msgs, res.messages   
-  end
-
-  def test_empty_observer_raises
-    scheduler = Rx::TestScheduler.new
-
-    xs = Rx::Observable.empty(scheduler)
-
-    observer = Rx::Observer.configure do |obs|
-      obs.on_completed { raise RuntimeError.new }
-    end
-
-    xs.subscribe observer
-
-    assert_raises(RuntimeError) { scheduler.start }
-  end
-
   # Never methods
 
   def test_never_basic
